[REFACTOR] move MSstatsFile to separate class
