+++ src/openms/source/FORMAT/sources.cmake
@@ -38,6 +38,7 @@ MS2File.cpp
 MSNumpressCoder.cpp
 MSPFile.cpp
 MSPGenericFile.cpp
+MSstatsFile.cpp
 MascotInfile.cpp
 MascotGenericFile.cpp
 MascotRemoteQuery.cpp
